# docs/api/openapi.yaml
openapi: 3.0.1
info:
  title: WeJam API
  version: v1
servers:
- description: Local development
  url: http://localhost:8080
security:
- bearerAuth: []
paths:
  /api/v1/admin/actions:
    get:
      operationId: adminActions
      parameters:
      - in: query
        name: before
        required: false
        schema:
          type: string
          format: date-time
      - in: query
        name: beforeId
        required: false
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminActionPage"
          description: OK
      tags:
      - Admin
  /api/v1/admin/venue-verifications:
    get:
      operationId: venueVerificationQueue
      parameters:
      - in: query
        name: status
        required: false
        schema:
          type: string
          default: PENDING
          enum:
          - PENDING
          - VERIFIED
          - REJECTED
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VenueVerificationItem"
          description: OK
      tags:
      - Admin
  /api/v1/admin/venues/{venueId}:
    get:
      operationId: reviewVenue
      parameters:
      - in: path
        name: venueId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenueReviewResponse"
          description: OK
      tags:
      - Admin
  /api/v1/admin/venues/{venueId}/approve:
    post:
      operationId: approveVenue
      parameters:
      - in: path
        name: venueId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenueResponse"
          description: OK
      tags:
      - Admin
  /api/v1/admin/venues/{venueId}/reject:
    post:
      operationId: rejectVenue
      parameters:
      - in: path
        name: venueId
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectVerificationRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenueResponse"
          description: OK
      tags:
      - Admin
  /api/v1/auth/token:
    post:
      operationId: exchange
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
          description: OK
      security: []
      tags:
      - Auth
  /api/v1/host-profiles/{profileId}:
    get:
      operationId: getHostProfile
      parameters:
      - in: path
        name: profileId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HostProfileResponse"
          description: OK
      tags:
      - Hosts
  /api/v1/me:
    get:
      operationId: me
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
          description: OK
      tags:
      - Me
    patch:
      operationId: updateProfile
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
          description: OK
      tags:
      - Me
  /api/v1/me/host-invites:
    get:
      operationId: myHostInvites
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HostInviteResponse"
          description: OK
      tags:
      - Hosts
  /api/v1/me/host-invites/{hostProfileId}/accept:
    post:
      operationId: acceptHostInvite
      parameters:
      - in: path
        name: hostProfileId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: No Content
      tags:
      - Hosts
  /api/v1/me/host-invites/{hostProfileId}/decline:
    post:
      operationId: declineHostInvite
      parameters:
      - in: path
        name: hostProfileId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: No Content
      tags:
      - Hosts
  /api/v1/me/host-memberships/{hostProfileId}:
    delete:
      operationId: leaveHostGroup
      parameters:
      - in: path
        name: hostProfileId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: No Content
      tags:
      - Hosts
  /api/v1/me/host-profile:
    get:
      operationId: getMyHostProfile
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HostProfileResponse"
          description: OK
      tags:
      - Hosts
    put:
      operationId: saveMyHostProfile
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HostProfileRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HostProfileResponse"
          description: OK
      tags:
      - Hosts
  /api/v1/me/host-profile/invites:
    get:
      operationId: pendingInvites
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PendingInviteResponse"
          description: OK
      tags:
      - Hosts
    post:
      operationId: inviteMember
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteMemberRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PendingInviteResponse"
          description: Created
      tags:
      - Hosts
  /api/v1/me/host-profile/members/{userId}:
    delete:
      operationId: removeMember
      parameters:
      - in: path
        name: userId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: No Content
      tags:
      - Hosts
  /api/v1/me/roles:
    post:
      operationId: addRole
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddRoleRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
          description: OK
      tags:
      - Me
  /api/v1/me/venues:
    get:
      operationId: myVenues
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VenueSummary"
          description: OK
      tags:
      - Venues
  /api/v1/venues:
    post:
      operationId: createVenue
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VenueRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenueResponse"
          description: Created
      tags:
      - Venues
  /api/v1/venues/{venueId}:
    delete:
      operationId: deleteVenue
      parameters:
      - in: path
        name: venueId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: No Content
      tags:
      - Venues
    get:
      operationId: getVenue
      parameters:
      - in: path
        name: venueId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenueResponse"
          description: OK
      tags:
      - Venues
    put:
      operationId: updateVenue
      parameters:
      - in: path
        name: venueId
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VenueRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenueResponse"
          description: OK
      tags:
      - Venues
  /api/v1/venues/{venueId}/resubmit:
    post:
      operationId: resubmitVenue
      parameters:
      - in: path
        name: venueId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenueResponse"
          description: OK
      tags:
      - Venues
  /api/v1/venues/{venueId}/spaces:
    post:
      operationId: createSpace
      parameters:
      - in: path
        name: venueId
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpaceRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpaceResponse"
          description: Created
      tags:
      - Venues
  /api/v1/venues/{venueId}/spaces/{spaceId}:
    delete:
      operationId: deleteSpace
      parameters:
      - in: path
        name: venueId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: spaceId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: No Content
      tags:
      - Venues
    put:
      operationId: updateSpace
      parameters:
      - in: path
        name: venueId
        required: true
        schema:
          type: string
          format: uuid
      - in: path
        name: spaceId
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpaceRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpaceResponse"
          description: OK
      tags:
      - Venues
components:
  schemas:
    AddRoleRequest:
      type: object
      properties:
        role:
          type: string
          description: Only HOST or VENUE_ADMIN are accepted
          enum:
          - USER
          - HOST
          - VENUE_ADMIN
          - PLATFORM_ADMIN
      required:
      - role
    AdminActionPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminActionResponse"
        nextBefore:
          type: string
          format: date-time
          nullable: true
        nextBeforeId:
          type: string
          format: uuid
          nullable: true
      required:
      - items
      - nextBefore
      - nextBeforeId
    AdminActionResponse:
      type: object
      properties:
        action:
          type: string
          enum:
          - VENUE_APPROVED
          - VENUE_REJECTED
        adminId:
          type: string
          format: uuid
        adminName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        id:
          type: string
          format: uuid
        reason:
          type: string
          nullable: true
        targetId:
          type: string
          format: uuid
        targetType:
          type: string
          enum:
          - VENUE
      required:
      - action
      - adminId
      - adminName
      - createdAt
      - id
      - reason
      - targetId
      - targetType
    HostInviteResponse:
      type: object
      properties:
        groupKind:
          type: string
          enum:
          - BAND
          - FRIENDS
          - COMMUNITY
        groupName:
          type: string
        hostProfileId:
          type: string
          format: uuid
        invitedAt:
          type: string
          format: date-time
        invitedByName:
          type: string
          nullable: true
      required:
      - groupKind
      - groupName
      - hostProfileId
      - invitedAt
      - invitedByName
    HostMemberDto:
      type: object
      description: "Owner first (admin), then accepted members. Pending invites are\
        \ never listed here."
      properties:
        admin:
          type: boolean
          description: The profile owner
        displayName:
          type: string
          nullable: true
        userId:
          type: string
          format: uuid
      required:
      - admin
      - displayName
      - userId
    HostProfileRequest:
      type: object
      properties:
        area:
          type: string
          example: Indiranagar
          maxLength: 80
          minLength: 0
          nullable: true
        bio:
          type: string
          example: Four friends running slow blues nights in Indiranagar since 2024.
          maxLength: 200
          minLength: 0
          nullable: true
        genres:
          type: array
          items:
            type: string
            enum:
            - BLUES
            - JAZZ
            - ROCK
            - INDIE
            - POP
            - FOLK
            - SOUL
            - FUNK
            - SUFI
            - BOLLYWOOD
            - INDIAN_CLASSICAL
            - WESTERN_CLASSICAL
            - FUSION
            - HIP_HOP
            - ELECTRONIC
            - METAL
            - ACOUSTIC_COVERS
          maxItems: 5
          minItems: 0
          uniqueItems: true
        groupKind:
          type: string
          description: Required when type is GROUP
          enum:
          - BAND
          - FRIENDS
          - COMMUNITY
          nullable: true
        groupName:
          type: string
          description: Required when type is GROUP
          example: The Low Notes
          maxLength: 80
          minLength: 0
          nullable: true
        instagramHandle:
          type: string
          description: "Instagram username, with or without @"
          example: thelownotes
          nullable: true
          pattern: "@?[A-Za-z0-9._]{1,30}"
        mediaLinks:
          type: array
          items:
            $ref: "#/components/schemas/MediaLinkDto"
          maxItems: 5
          minItems: 0
        type:
          type: string
          enum:
          - INDIVIDUAL
          - GROUP
      required:
      - genres
      - mediaLinks
      - type
    HostProfileResponse:
      type: object
      properties:
        area:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        displayName:
          type: string
          description: "Group name, or the owner's display name for an individual\
            \ (null if not set yet)"
          nullable: true
        genres:
          type: array
          items:
            type: string
            enum:
            - BLUES
            - JAZZ
            - ROCK
            - INDIE
            - POP
            - FOLK
            - SOUL
            - FUNK
            - SUFI
            - BOLLYWOOD
            - INDIAN_CLASSICAL
            - WESTERN_CLASSICAL
            - FUSION
            - HIP_HOP
            - ELECTRONIC
            - METAL
            - ACOUSTIC_COVERS
        groupKind:
          type: string
          enum:
          - BAND
          - FRIENDS
          - COMMUNITY
          nullable: true
        id:
          type: string
          format: uuid
        instagramHandle:
          type: string
          nullable: true
        mediaLinks:
          type: array
          items:
            $ref: "#/components/schemas/MediaLinkDto"
        members:
          type: array
          description: "Owner first (admin), then accepted members. Pending invites\
            \ are never listed here."
          items:
            $ref: "#/components/schemas/HostMemberDto"
        type:
          type: string
          enum:
          - INDIVIDUAL
          - GROUP
      required:
      - area
      - bio
      - displayName
      - genres
      - groupKind
      - id
      - instagramHandle
      - mediaLinks
      - members
      - type
    InviteMemberRequest:
      type: object
      properties:
        phoneNumber:
          type: string
          example: "+911234567890"
          minLength: 1
          pattern: "\\+[1-9]\\d{7,14}"
      required:
      - phoneNumber
    MeResponse:
      type: object
      properties:
        displayName:
          type: string
          description: Null until the user completes their profile
          nullable: true
        id:
          type: string
          format: uuid
        phone:
          type: string
          description: E.164; null for non-phone sign-in methods
          nullable: true
        roles:
          type: array
          items:
            type: string
            enum:
            - USER
            - HOST
            - VENUE_ADMIN
            - PLATFORM_ADMIN
          uniqueItems: true
      required:
      - displayName
      - id
      - phone
      - roles
    MediaLinkDto:
      type: object
      properties:
        title:
          type: string
          example: Live at Amber Room
          maxLength: 80
          minLength: 0
          nullable: true
        url:
          type: string
          example: https://youtu.be/abc123
          maxLength: 300
          minLength: 0
          pattern: https?://\S+
      required:
      - title
      - url
    PendingInviteResponse:
      type: object
      properties:
        invitedAt:
          type: string
          format: date-time
        maskedPhone:
          type: string
          example: +91 12•••••890
        userId:
          type: string
          format: uuid
      required:
      - invitedAt
      - maskedPhone
      - userId
    RejectVerificationRequest:
      type: object
      properties:
        reason:
          type: string
          description: Sent verbatim to the venue
          example: The FSSAI number doesn't match the registry. Please check the 14
            digits on your license.
          maxLength: 500
          minLength: 0
      required:
      - reason
    SpaceRequest:
      type: object
      properties:
        capacity:
          type: integer
          format: int32
          example: 60
          maximum: 1000
          minimum: 1
        houseRules:
          type: string
          description: Shown on every event at this space
          example: 18+ after 9 PM. Min. spend ₹300 for audience.
          maxLength: 1000
          minLength: 0
          nullable: true
        name:
          type: string
          example: Main floor
          maxLength: 80
          minLength: 0
        soundCurfew:
          type: string
          description: Local time (venue time zone) events here must end by; null
            = no curfew
          example: 23:00
          nullable: true
        soundPolicy:
          type: string
          enum:
          - ACOUSTIC_ONLY
          - AMPLIFIED_ALLOWED
      required:
      - capacity
      - name
      - soundPolicy
    SpaceResponse:
      type: object
      properties:
        capacity:
          type: integer
          format: int32
        houseRules:
          type: string
          nullable: true
        id:
          type: string
          format: uuid
        name:
          type: string
        soundCurfew:
          type: string
          example: 23:00
          nullable: true
        soundPolicy:
          type: string
          enum:
          - ACOUSTIC_ONLY
          - AMPLIFIED_ALLOWED
      required:
      - capacity
      - houseRules
      - id
      - name
      - soundCurfew
      - soundPolicy
    TokenRequest:
      type: object
      properties:
        firebaseIdToken:
          type: string
          minLength: 1
      required:
      - firebaseIdToken
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        expiresIn:
          type: integer
          format: int64
          description: Seconds until accessToken expires
        tokenType:
          type: string
          example: Bearer
      required:
      - accessToken
      - expiresIn
      - tokenType
    UpdateProfileRequest:
      type: object
      properties:
        displayName:
          type: string
          maxLength: 50
          minLength: 0
      required:
      - displayName
    VenueRequest:
      type: object
      properties:
        addressLine:
          type: string
          example: "12th Main, HAL 2nd Stage"
          maxLength: 200
          minLength: 0
        city:
          type: string
          example: Bengaluru
          maxLength: 80
          minLength: 0
        description:
          type: string
          maxLength: 1000
          minLength: 0
          nullable: true
        fssaiNumber:
          type: string
          description: FSSAI license number. Changing it resets verification to PENDING.
          example: "12345678901234"
          pattern: "\\d{14}"
        hostingMode:
          type: string
          enum:
          - OPEN
          - SELF_ONLY
        latitude:
          type: number
          format: double
          example: 12.9784
          exclusiveMaximum: false
          exclusiveMinimum: false
          maximum: 90
          minimum: -90
        longitude:
          type: number
          format: double
          example: 77.6408
          exclusiveMaximum: false
          exclusiveMinimum: false
          maximum: 180
          minimum: -180
        name:
          type: string
          example: Amber Room
          maxLength: 80
          minLength: 0
      required:
      - addressLine
      - city
      - fssaiNumber
      - hostingMode
      - latitude
      - longitude
      - name
    VenueResponse:
      type: object
      properties:
        addressLine:
          type: string
        city:
          type: string
        description:
          type: string
          nullable: true
        fssaiNumber:
          type: string
        hostingMode:
          type: string
          enum:
          - OPEN
          - SELF_ONLY
        id:
          type: string
          format: uuid
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        name:
          type: string
        rejectionReason:
          type: string
          nullable: true
        spaces:
          type: array
          items:
            $ref: "#/components/schemas/SpaceResponse"
        verificationStatus:
          type: string
          enum:
          - PENDING
          - VERIFIED
          - REJECTED
      required:
      - addressLine
      - city
      - description
      - fssaiNumber
      - hostingMode
      - id
      - latitude
      - longitude
      - name
      - rejectionReason
      - spaces
      - verificationStatus
    VenueReviewResponse:
      type: object
      properties:
        checks:
          $ref: "#/components/schemas/VerificationChecks"
        ownerName:
          type: string
          nullable: true
        ownerPhone:
          type: string
          nullable: true
        venue:
          $ref: "#/components/schemas/VenueResponse"
      required:
      - checks
      - ownerName
      - ownerPhone
      - venue
    VenueSummary:
      type: object
      properties:
        city:
          type: string
        id:
          type: string
          format: uuid
        name:
          type: string
        spaceCount:
          type: integer
          format: int64
        verificationStatus:
          type: string
          enum:
          - PENDING
          - VERIFIED
          - REJECTED
      required:
      - city
      - id
      - name
      - spaceCount
      - verificationStatus
    VenueVerificationItem:
      type: object
      properties:
        checks:
          $ref: "#/components/schemas/VerificationChecks"
        city:
          type: string
        fssaiNumber:
          type: string
        name:
          type: string
        ownerName:
          type: string
          nullable: true
        requestedAt:
          type: string
          format: date-time
        status:
          type: string
          enum:
          - PENDING
          - VERIFIED
          - REJECTED
        venueId:
          type: string
          format: uuid
      required:
      - checks
      - city
      - fssaiNumber
      - name
      - ownerName
      - requestedAt
      - status
      - venueId
    VerificationChecks:
      type: object
      properties:
        fssaiStructureLooksValid:
          type: boolean
          description: "14 digits, licence type 1/2, state code 01–38 (offline heuristic,\
            \ not the registry)"
        otherVenuesWithSameFssai:
          type: integer
          format: int64
          description: Other venues using the same FSSAI number
      required:
      - fssaiStructureLooksValid
      - otherVenuesWithSameFssai
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http